requirements.txt
Flask==2.3.3
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3
requests==2.31.0
poppler-utils==23.01.0
python-dotenv==1.0.0